Add asynchronously rendered example
The latest version of the happo.io client will wait for content to be
available before we complete the rendering of an example. It also
correctly fetches styles injected via style.sheet.insertRule. I wanted
to put these new features to a test in a real project, so here goes.